search, upd8: share schema between client and upd8 code
