+++ src/util/searchSchema.js
@@ -1,0 +1,22 @@
+// Index structures shared by client and server.
+
+export function makeSearchIndexes(FlexSearch) {
+  const indexes = {
+    albums: new FlexSearch.Document({
+      id: "reference",
+      index: ["name", "groups"],
+    }),
+
+    tracks: new FlexSearch.Document({
+      id: "reference",
+      index: ["name", "album", "artists", "directory", "additionalNames"],
+    }),
+
+    artists: new FlexSearch.Document({
+      id: "reference",
+      index: ["names"],
+    }),
+  };
+
+  return indexes;
+}
